Changed structure of tasks and created new scripts
publish-results.sh split to:
- post-build-create-image-manifest.sh
- post-build-sort-out-downloads.sh

Second one also takes care of downloads/ dir so we do not push the same
archives again and again to snapshots.linaro.org

Cleaning of disk space got moved from Jenkins to pre-build-do-cleanup.sh
script.

pre-build-root-install-dependencies.sh just got renamed so it is visible
when it is run

init-and-build.sh also got renamed as it not only initialize but also
do a build.

Signed-off-by: Marcin Juszkiewicz <marcin.juszkiewicz@linaro.org>