Add support for manifest groups options
Advanced manifest can use 'groups' to tie projects together, such
that only a subset of the project are fetched/sync'd when running
repo sync.

The 'groups' is provided to repo init command.

For repo manifest documentation:

=
Attribute `groups`: List of groups to which this project belongs,
whitespace or comma separated.  All projects belong to the group
"all", and each project automatically belongs to a group of
its name:`name` and path:`path`.  E.g. for
<project name="monkeys" path="barrel-of"/>, that project
definition is implicitly in the following manifest groups:
default, name:monkeys, and path:barrel-of.  If you place a project in the
group "notdefault", it will not be automatically downloaded by repo.
If the project has a parent element, the `name` and `path` here
are the prefixed ones.
=

Since '-g' option to init-and-build.sh was used already, '-m' is
used...

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>