create a dedicated function to initialize the 'build' folder
The main idea here is to start using jenkins-setup scripts as generic
tools that facilitates the setup, configuration and builds for Linaro
OE builds, and to remove any configuration which is specific to a
particular build. When Linaro OE CI started there was only one build,
we now want to reuse the setup tools for every Linaro OE builds. So
only settings which are related to Linaro CI should persist
eventually.

So this commit isolates the 'build environemnt initiliazation' in its
own function, so that it can later be replaced and parameterized.

Change-Id: If347103654339d42a027580eed2f5e592b8286f1
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>