set Linaro CI site.conf specific options in a standalone function
This commit splits the Linaro CI (Jenkins) options into a specific
function, so that these settings can be done even when using a custom
workspace init function. e.g. after the build specific 'init env'
script/function is done. It is important to note that
init_env_linaro_ci() will be called for every build, so this function
must only configured global configuration, not build-specific
attributes. Build specific options need to be configured in init_env
or in the custom init function.

Change-Id: Ifd3b04bfa60b6def780ea41f4b935ac186bbafe0
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>