post-build-create-image-manifest: fix path used by find command
In the original patch (b799530 ) was merged into master, we are using
'find . -name xxx'. As such we clearly depends on the value of $PWD,
and it was expected to be the in the proper tmp/deploy/image
folder. However the logic in that script had been rewritten, so the
original patch no longer works properly.  Since it is now executed
from the WORKSPACE, 'find' will search the entire WORKSPACE and get
some wrong files.

We need to restrict where 'find' searches files.

Change-Id: Id456f3061bbef52e7e69f20d4e79f0e57d6f052d
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>