functions.sh init-and-build.sh: add support for custom sstate dir folder
When doing builds for several OE versions (e.g. dylan, dora, master)
it is better to use a different sstate folder for each release,
otherwise the sstate-cache-management will delete older release
packages which is not a behavior that we need. This commit adds an
option to specify a 'subfolder' to use in the main sstate_dir area for
Linaro CI loop.

Change-Id: I49662c0ec6ebf577992a12240626ada999918c5a
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>