new script: srcrev-check.py
Since using ${AUTOREV} is consired harmful, we need to pin SRCREV
in our recipes. However, we still need a way to update SRCREV to keep
up to date with linaro CI.

This turns out to be a bit of mess, as the SRC_URI has a nonstandard
url format. To get the right SRCREV from right branch etc, we use
bitbakes "fetch" module, which isnt quite designed for use outside
bitbake.

Usage:

. oe-init-build-env ../build/
../jenkins-setup/srcrev-check.py -d ../meta-linaro/meta-linaro/recipes-kernel/

add the "-u" flag to actually update.

To be discussed - how should these updates be submitted new that we
have gerrit in between.

caveat: this script is only able to update recipes where git SRC_URI
and SRCREV are in same file.

Change-Id: I63ab1861dba767165d759685185c34ff19f6eda8
Signed-off-by: Riku Voipio <riku.voipio@linaro.org>